+++ src/persistence/core.go
@@ -1,0 +1,2 @@
+// Package persistence stores Git Town runstate on disk.
+package persistence
