+++ src/persistence/save_load_test.go
@@ -13,19 +13,8 @@ import (
 	"github.com/stretchr/testify/assert"
 )
 
-func TestSanitizePath(t *testing.T) {
+func TestLoadSave(t *testing.T) {
 	t.Parallel()
-	t.Run("SanitizePath", func(t *testing.T) {
-		t.Parallel()
-		tests := map[string]string{
-			"/home/user/development/git-town":        "home-user-development-git-town",
-			"c:\\Users\\user\\development\\git-town": "c-users-user-development-git-town",
-		}
-		for give, want := range tests {
-			have := persistence.SanitizePath(give)
-			assert.Equal(t, want, have)
-		}
-	})
 	t.Run("Save and Load", func(t *testing.T) {
 		t.Parallel()
 		runState := runstate.RunState{
