+++ src/runstate/core.go
@@ -1,6 +1,5 @@
-// Package runstate represents the current execution status of a Git Town command and allows to run or persist it to disk.
-// This is used by the "abort", "continue", and "undo" commands.
-// The central data structure is RunState.
+// Package runstate represents the current execution status of a Git Town command,
+// i.e. which steps to execute via the `runvm`.
 package runstate
 
 import (
