+++ src/runvm/core.go
@@ -1,0 +1,2 @@
+// Package runvm provides the virtual machine that executes steps of Git Town commands.
+package runvm
